# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gorilla Guardians API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /experiences:
    get:
      operationId: listExperiences
      tags: [experiences]
      summary: List experiences
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Experiences list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperienceListResponse"
    post:
      operationId: createExperience
      tags: [experiences]
      summary: Create experience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
  /experiences/{id}:
    get:
      operationId: getExperience
      tags: [experiences]
      summary: Get experience by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
    patch:
      operationId: updateExperience
      tags: [experiences]
      summary: Update experience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"

  # BOOKINGS
...
  /conversations:
    get:
      operationId: listConversations
      tags: [messages]
      summary: List conversations
      responses:
        "200":
          description: Conversations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [messages]
      summary: Create or get conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a conversation
      parameters:
        - name: conversationId
          in: query
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # FEEDBACK
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, admin, staff, artisan, customer]
        avatar:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        language:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, isActive, createdAt]
...
    Experience:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [guided_tour, homestay, workshop, dance_lesson, cooking_class]
        images:
          type: array
          items:
            type: string
        videoUrl:
          type: ["string", "null"]
        price:
          type: number
        duration:
          type: string
        capacity:
          type: integer
        difficultyLevel:
          type: ["string", "null"]
        includedItems:
          type: array
          items:
            type: string
        meetingPoint:
          type: ["string", "null"]
        cancellationPolicy:
          type: ["string", "null"]
        active:
          type: boolean
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, slug, description, type, images, price, duration, capacity, includedItems, active, reviewCount, createdAt]
    ExperienceListResponse:
      type: object
      properties:
        experiences:
          type: array
          items:
            $ref: "#/components/schemas/Experience"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [experiences, total, page, totalPages]
    ExperienceInput:
      type: object
      required: [title, description, type, price, duration, capacity]
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
        images:
          type: array
          items:
            type: string
        videoUrl:
          type: string
        price:
          type: number
        duration:
          type: string
        capacity:
          type: integer
        difficultyLevel:
          type: string
        includedItems:
          type: array
          items:
            type: string
        meetingPoint:
          type: string
        cancellationPolicy:
          type: string
        active:
          type: boolean
    ExperienceUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        duration:
          type: string
        capacity:
          type: integer
        images:
          type: array
          items:
            type: string
        includedItems:
          type: array
          items:
            type: string
        active:
          type: boolean

    # BOOKING
...
    Conversation:
      type: object
      properties:
        id:
          type: integer
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        lastMessage:
          type: ["string", "null"]
        unreadCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
      required: [id, participants, unreadCount, updatedAt]
    ConversationInput:
      type: object
      required: [participantId]
      properties:
        participantId:
          type: integer
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        fileUrl:
          type: ["string", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, conversationId, senderId, content, isRead, createdAt]
    MessageInput:
      type: object
      required: [conversationId, content]
      properties:
        conversationId:
          type: integer
        content:
          type: string
        fileUrl:
          type: string

    # FEEDBACK
...